feat(flight): double-precision FlightState position vector (#371)
FlightState::pos_world[3] is upgraded from float[3] to double[3],
closing the silent precision gap at large world-space distances.
At x = 1e5 m the float ULP (~0.0078 m) exceeds the per-step lateral
gravity displacement (~0.001 m); the position never moved. With
double, sub-millimetre increments accumulate correctly at any scale
including planet-radius distances needed by #354 and #355.

Interface changes (primary development — no compat constraints):
- IGravityField::accelWorld and geodeticAltitude params upgraded to
  const double[]; geodeticAltitude return upgraded to double
- CentralGravityField uses double arithmetic internally for the
  direction computation; float return via explicit cast (gravity
  magnitudes ~9.8 m/s² are fine in float)
- FlightIntegrator.cpp: two static_cast<float> at computeAtmosphere
  call sites (deliberate float boundary)
- WorldBroadcaster: setGroundElevationQuery callback upgraded to
  float(double,double); m_entityX/Z atomics upgraded to
  atomic<double> with static_assert(is_always_lock_free);
  cachedEntityX/Z() return double

Wire format unchanged (MsgEntityEntry::pos already double[3]).
kProtocolVersion stays at 1.

Adds precision regression tests in test_flight_integrator and
test_world_broadcaster that would fail against the old float[3].

+++ engine/flight/CentralGravityField.h
@@ -19,23 +19,23 @@ class CentralGravityField final : public IGravityField {
   public:
     explicit CentralGravityField(float R = 6'371'000.f, float g = 9.80665f) : m_GM(g * R * R), m_centerY(-R) {}
 
-    std::array<float, 3> accelWorld(const float p[3]) const override {
-        const float dx = p[0];
-        const float dy = p[1] - m_centerY; // p[1] - (-R) = p[1] + R
-        const float dz = p[2];
-        const float r2 = dx * dx + dy * dy + dz * dz;
-        if (r2 < 1.f)
+    std::array<float, 3> accelWorld(const double p[3]) const override {
+        const double dx = p[0];
+        const double dy = p[1] - static_cast<double>(m_centerY); // p[1] - (-R) = p[1] + R
+        const double dz = p[2];
+        const double r2 = dx * dx + dy * dy + dz * dz;
+        if (r2 < 1.0)
             return {0.f, 0.f, 0.f}; // at or inside planet centre — avoid divide-by-zero
-        const float r = std::sqrt(r2);
-        const float a = -m_GM / r2; // magnitude (negative = toward centre)
-        return {a * dx / r, a * dy / r, a * dz / r};
+        const double r = std::sqrt(r2);
+        const double a = -static_cast<double>(m_GM) / r2; // magnitude (negative = toward centre)
+        return {static_cast<float>(a * dx / r), static_cast<float>(a * dy / r), static_cast<float>(a * dz / r)};
     }
 
-    float geodeticAltitude(const float p[3]) const override {
-        const float dx = p[0];
-        const float dy = p[1] - m_centerY;
-        const float dz = p[2];
-        return std::sqrt(dx * dx + dy * dy + dz * dz) + m_centerY; // |r_vec| - R
+    double geodeticAltitude(const double p[3]) const override {
+        const double dx = p[0];
+        const double dy = p[1] - static_cast<double>(m_centerY);
+        const double dz = p[2];
+        return std::sqrt(dx * dx + dy * dy + dz * dz) + static_cast<double>(m_centerY); // |r_vec| - R
     }
 
     // Pre-built Earth instance (R = 6 371 000 m, g = 9.80665 m/s²).
